+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Version 3.10-SNAPSHOT
 ADDED:
 
 - new extension class for resolve the free memory and the memory status and freeing memory that is available for the application by running the garbage collector 
-  
+- new method created that set the current thread priority
+
 Version 3.9
 -------------